+++ src/security/noise/handshake_message_marshaller_impl.cpp
@@ -59,7 +59,8 @@ namespace libp2p::security::noise {
       const HandshakeMessage &msg) const {
     OUTCOME_TRY(proto_msg, handyToProto(msg));
     common::ByteArray out_msg(proto_msg.ByteSizeLong());
-    if (not proto_msg.SerializeToArray(out_msg.data(), out_msg.size())) {
+    if (not proto_msg.SerializeToArray(out_msg.data(),
+                                       static_cast<int>(out_msg.size()))) {
       return Error::MESSAGE_SERIALIZING_ERROR;
     }
     return out_msg;
@@ -69,7 +70,8 @@ namespace libp2p::security::noise {
   HandshakeMessageMarshallerImpl::unmarshal(
       gsl::span<const uint8_t> msg_bytes) const {
     protobuf::NoiseHandshakePayload proto_msg;
-    if (not proto_msg.ParseFromArray(msg_bytes.data(), msg_bytes.size())) {
+    if (not proto_msg.ParseFromArray(msg_bytes.data(),
+                                     static_cast<int>(msg_bytes.size()))) {
       return Error::MESSAGE_DESERIALIZING_ERROR;
     }
     return protoToHandy(proto_msg);
